Lets add a cmake build system
The build is done on Linux, so I needed to convert to includes
to upper case. I also fixed some issues found when compiling in
GCC. Clang "should be OK", not tested yet, but I will test.

CMake targets a program, so it will fails to link (the code
is originally a DLL, so a library will be the target
eventually).